Booking functionality & User bookings page
When clicked on Book this place button, two inputs will be added (name and mobile phone). Name will be automatically filled, as the name of the user that has already been provided. The booked place will be displayed in my bookings section, including the title of the place, the number of nights (chosen using calendars), the exact dates and the total price (number of nights multiplied with a price per night).